rename DSP.{c,h} to dsp.{c,h}
